# apps/oai-backend-py/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAI-Compatible Backend API
  description: |
    OpenAI-compatible API proxy for Ollama with additional features for document processing,
    vector database management, and tool calling.
    
    This API provides:
    - OpenAI-compatible chat completions and embeddings endpoints
    - OAuth2 authentication via Keycloak/GitLab
    - Document upload and processing pipelines
    - Milvus vector database collection management
    - Direct tool calling capabilities
  version: 0.1.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /v1/chat/completions:
    post:
      tags:
        - OpenAI Compatible
      summary: Create chat completion
      description: |
        OpenAI-compatible chat completions endpoint with streaming support and tool calling.
        Supports both streaming and non-streaming modes. Automatically handles tool calls
        by executing them and re-querying the LLM until a final response is generated.
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Chat completion response (non-streaming) or stream of chunks (streaming)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletion'
            text/event-stream:
              schema:
                type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/embeddings:
    post:
      tags:
        - OpenAI Compatible
      summary: Create embeddings
      description: |
        OpenAI-compatible embeddings endpoint. Generates vector embeddings for text input
        using the specified model.
      operationId: createEmbedding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingRequest'
      responses:
        '200':
          description: Embedding response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/models:
    get:
      tags:
        - OpenAI Compatible
      summary: List models
      description: |
        Returns a list of available models from Ollama and custom agents,
        formatted in OpenAI-compatible format.
      operationId: listModels
      responses:
        '200':
          description: List of available models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelList'
        '502':
          description: Failed to fetch models from Ollama
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/collections:
    get:
      tags:
        - Collections
      summary: List collections
      description: |
        Returns all Milvus collections with their metadata.
        Requires authentication with a valid JWT token containing Milvus credentials.
      operationId: listCollections
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of collections with metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionsResponse'
        '401':
          description: Unauthorized - Milvus token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to list collections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Collections
      summary: Create collection
      description: |
        Creates a new Milvus collection from a full CrawlerConfig JSON.
        Collection name is taken from crawler_config.database.collection.
        When access_level is group_only, access_groups must be non-empty and is used for security_groups and role grants.
        Requires authentication with a valid JWT token containing Milvus credentials.
      operationId: createCollection
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
      responses:
        '200':
          description: Collection created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCollectionResponse'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Milvus token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Collection already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to create collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/pipelines/{name}:
    get:
      tags:
        - Collections
      summary: Get pipeline config by name
      description: |
        Returns the full crawler-config JSON for the given pipeline template.
        No authentication required.
      operationId: getPipelineConfig
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Pipeline template name (e.g. standard, academic)
      responses:
        '200':
          description: Full crawler config object
          content:
            application/json:
              schema:
                type: object
                description: Crawler config JSON (database, embeddings, llm, etc.)
        '404':
          description: Pipeline not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/users:
    get:
      tags:
        - Collections
      summary: List users
      description: |
        Returns all Milvus users with their roles.
        Requires authentication with a valid JWT token.
      operationId: listUsers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users with roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        '401':
          description: Unauthorized - Milvus token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to list users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/roles:
    get:
      tags:
        - Collections
      summary: List roles
      description: |
        Returns all Milvus roles with their privileges.
        Requires authentication with a valid JWT token.
      operationId: listRoles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of roles with privileges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesResponse'
        '401':
          description: Unauthorized - Milvus token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to list roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/collections/{collection_name}/process:
    post:
      tags:
        - Documents
      summary: Process document to extract metadata
      description: |
        Processes a document to extract metadata without uploading to the database.
        Uses the collection's pipeline configuration.
      operationId: processDocumentForCollection
      security:
        - BearerAuth: []
      parameters:
        - name: collection_name
          in: path
          required: true
          schema:
            type: string
          description: Name of the collection (used to load config)
          examples:
            - my_collection
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file to process (PDF, etc.)
      responses:
        '200':
          description: Document metadata extracted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessedDocument'
        '401':
          description: Unauthorized - Milvus token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to process document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/collections/{collection_name}/upload:
    post:
      tags:
        - Documents
      summary: Upload document to collection
      description: |
        Uploads and processes a document to a collection using the collection's
        pipeline configuration. Either file or markdown_content must be provided.
      operationId: uploadDocumentToCollection
      security:
        - BearerAuth: []
      parameters:
        - name: collection_name
          in: path
          required: true
          schema:
            type: string
          description: Name of the collection to upload to
          examples:
            - my_collection
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file to upload (PDF, etc.). Optional if markdown_content is provided.
                markdown_content:
                  type: string
                  description: Pre-converted markdown (skips conversion). Optional if file is provided.
                metadata_override:
                  type: string
                  description: |
                    JSON string with pre-extracted metadata (skips extraction).
                    Example: {"title":"My Doc","author":["John"]}
                security_groups:
                  type: string
                  description: |
                    JSON string with security groups for the document.
                    Example: ["public","team_a"]
      responses:
        '200':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Bad request - invalid metadata_override/security_groups JSON or missing both file and markdown_content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Milvus token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to upload document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service unavailable - failed to connect to vector database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/search:
    post:
      tags:
        - Search
      summary: Search collection
      description: |
        Hybrid search over a Milvus collection. Requires authentication with
        a valid JWT token containing Milvus credentials.
      operationId: search
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Milvus token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Search failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/agents/{agent}:
    post:
      tags:
        - OpenAI Compatible
      summary: Agentic RAG completion
      description: |
        Agentic endpoint invoked by agent name. Requires authentication.
      operationId: createAgentCompletion
      security:
        - BearerAuth: []
      parameters:
        - name: agent
          in: path
          required: true
          schema:
            type: string
          description: Agent name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRequest'
      responses:
        '200':
          description: Chat completion (non-streaming) or server-sent events (streaming)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletion'
            text/event-stream:
              schema:
                type: string
        '400':
          description: Bad request - missing messages or invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/tools:
    post:
      tags:
        - Tools
      summary: Call tool directly
      description: |
        Executes a tool directly without going through chat completions.
        Available tools:
        - `get_weather`: Get weather information for a location
        - `search`: Search the Milvus vector database
        
        Metadata can be provided for user-specific tool initialization.
      operationId: callTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolCallRequest'
      responses:
        '200':
          description: Tool execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolCallResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Tool not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Tool execution error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from OAuth2 authentication flow.
        Token should be included in the Authorization header as: `Bearer <token>`
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
      required:
        - detail
    ChatCompletionRequest:
      type: object
      required:
        - model
        - messages
      properties:
        model:
          type: string
          description: Model identifier (e.g., "llama3.2:1b", "radchat")
          examples:
            - llama3.2:1b
        messages:
          type: array
          description: List of messages in the conversation
          items:
            $ref: '#/components/schemas/Message'
        stream:
          type: boolean
          description: Whether to stream the response
          default: false
        temperature:
          type: number
          description: Sampling temperature (0-2)
          minimum: 0
          maximum: 2
        max_tokens:
          type: integer
          description: Maximum tokens to generate
        tools:
          type: array
          description: List of tools available to the model (auto-added if not provided)
          items:
            type: object
        tool_choice:
          type: string
          description: Tool choice strategy
          enum:
            - none
            - auto
            - required
...
    ChatCompletion:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          examples:
            - chat.completion
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
        usage:
          type: object
        history:
          type: array
          description: Full call history including internal agent steps and tool calls
          items:
            $ref: '#/components/schemas/Message'
    EmbeddingRequest:
      type: object
      required:
        - model
        - input
      properties:
        model:
          type: string
          description: Embedding model identifier
          examples:
            - all-minilm:v2
        input:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Text or list of texts to embed
    EmbeddingResponse:
      type: object
      properties:
        object:
          type: string
          examples:
            - list
        data:
          type: array
          items:
            type: object
            properties:
              object:
                type: string
                examples:
                  - embedding
              embedding:
                type: array
                items:
                  type: number
              index:
                type: integer
        model:
          type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            total_tokens:
              type: integer
    ModelList:
      type: object
      properties:
        object:
          type: string
          examples:
            - list
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              object:
                type: string
                examples:
                  - model
              created:
                type: integer
              owned_by:
                type: string
...
    CollectionsResponse:
      type: object
      properties:
        collection_names:
          type: array
          items:
            type: string
          description: List of collection names
        collections:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CollectionInfo'
          description: Metadata for each collection by name
    UsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: User identifier
              name:
                type: string
                description: User name
              roles:
                type: array
                items:
                  type: string
                description: List of role names assigned to the user
            required:
              - id
              - name
              - roles
          description: List of users with their roles
    RolesResponse:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
          description: List of role names
    CreateCollectionRequest:
      type: object
      required:
        - crawler_config
      properties:
        access_level:
          type: string
          enum:
            - public
            - private
            - group_only
            - admin
          default: public
          description: Access level for the collection
        access_groups:
          type: array
          items:
            type: string
          default: []
          description: List of role/group names to grant read access; required when access_level is group_only
        crawler_config:
          type: object
          description: Full CrawlerConfig JSON; must include database.collection
    CreateCollectionResponse:
      type: object
      properties:
        collection_name:
          type: string
          description: Name of the created collection
        message:
          type: string
          examples:
            - Collection created successfully
        roles:
          type: array
          items:
            type: string
          description: Roles granted read access to the collection
    ProcessedDocument:
      type: object
      properties:
        metadata:
          type: object
          description: Extracted document metadata
          additionalProperties: true
        markdown_content:
          type: string
          description: Converted markdown content
        file_name:
          type: string
          description: Name of the processed file
        file_size:
          type: [integer, "null"]
          description: Size of the file in bytes
    UploadResponse:
      type: object
      properties:
        message:
          type: string
          examples:
            - Document processed and uploaded successfully
        collection_name:
          type: string
          description: Name of the collection the document was uploaded to
        document_id:
          type: [string, "null"]
          description: Identifier for the processed document
        chunks_created:
          type: integer
          description: Number of chunks created
          default: 0
        processing_time_sec:
          type: number
          description: Time taken to process the document in seconds
    ToolCallRequest:
      type: object
      required:
        - name
        - arguments
      properties:
        name:
          type: string
          description: Name of the tool to execute
          examples:
            - get_weather
        arguments:
          type: object
          description: Arguments for the tool
          additionalProperties: true
          examples:
            - location: San Francisco, CA
        metadata:
          type: object
          description: Optional metadata for user-specific tool initialization
          additionalProperties: true
          examples:
            - user_id: user123
...
    SearchRequest:
      type: object
      required:
        - collection
      properties:
        collection:
          type: string
          minLength: 1
          description: Name of the Milvus collection to search
        text:
          type: string
          default: ""
          description: Query text to search for
        filters:
          type: array
          items:
            type: string
          default: []
          description: List of filter expressions (e.g., title == \"example\")
        limit:
          type: integer
          minimum: 1
          maximum: 10000
          default: 100
          description: Maximum number of results to return
...
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
          description: List of search hits
        total:
          type: integer
          description: Total number of results
    AgentRequest:
      type: object
      required:
        - messages
      properties:
        model:
          type: string
          description: Underlying LLM model (defaults to config)
        token:
          type: string
          description: Milvus token (optional when authenticated)
        collection:
          type: string
          minLength: 1
          description: Name of the Milvus collection to connect to
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: Conversation messages
        stream:
          type: boolean
          default: false
          description: Whether to stream the response